Add `ZSH_LS_PREFER_LS` as a config environment variable
This is to enable GNU ls configs even if exa is installed. Currently no
easy way to do it because we must uninstall exa or remove directories
contains exa from $PATH. But they are bothersome. So add an environment
variable to prioritize the configs for an easy configuration way. It is
easy and major way to config plugin.